Fix segfault on object destruction
The clear_and_dispose operation on intrusive lists seemingly would
disconnect the nodes from the list first prior to running the disposer
on them; this meant the object_list_hook destructor would then attempt
to double-unlink itself and segfault on dereferencing a null pair of
next/prev pointers.

+++ src/object_list_hook.hpp
@@ -38,6 +38,8 @@ namespace meorawr::hyjal {
     template<typename T>
     object_list_hook<T>::~object_list_hook() noexcept
     {
-        node_algorithms::unlink(this);
+        if (!node_algorithms::inited(this)) {
+            node_algorithms::unlink(this);
+        }
     }
 }
